+++ src/services/absences/index.js
@@ -1,14 +1,29 @@
 import absences from '../../support/absences';
+import moment from 'moment';
 
-export const list = () => {
-  return absences.payload;
-};
+export const list = () =>
+  absences.payload;
+
+/* startDate and endDate have to be passed as moment objects,
+   to display absences in a calendar view
+   canMove: false => to not move this event in a calendar
+*/
+export const listWithMoment = () =>
+  list().map((absence) => {
+    return {
+      ...absence,
+      startDate: moment(absence.startDate),
+      endDate: moment(absence.endDate).add(24, 'hour'),
+      canMove: false
+    };
+  });
 
 export const get = (id) =>
-  list().find((absence) => absence.id === id)
+  listWithMoment().find((absence) => absence.id === id)
   || null;
 
 export default {
   get,
-  list
+  list,
+  listWithMoment
 };